chore: rename src/ckb_chain to src/ckb
