+++ src/errors.rs
@@ -3,7 +3,7 @@ use tentacle::{error::SendErrorKind, secio::PeerId};
 use thiserror::Error;
 
 use crate::{
-    ckb_chain::FundingError,
+    ckb::FundingError,
     fiber::{
         channel::{ChannelActorMessage, ProcessingChannelError},
         types::Hash256,
