+++ src/main.rs
@@ -1,5 +1,5 @@
 use cfn_node::cch::CchMessage;
-use cfn_node::ckb_chain::contracts::init_contracts_context;
+use cfn_node::ckb::contracts::init_contracts_context;
 use cfn_node::store::Store;
 use ractor::Actor;
 use tentacle::multiaddr::Multiaddr;
@@ -12,7 +12,7 @@ use tracing_subscriber::{fmt, EnvFilter};
 use std::str::FromStr;
 
 use cfn_node::actors::RootActor;
-use cfn_node::ckb_chain::CkbChainActor;
+use cfn_node::ckb::CkbChainActor;
 use cfn_node::fiber::{channel::ChannelSubscribers, NetworkActorCommand, NetworkActorMessage};
 use cfn_node::tasks::{
     cancel_tasks_and_wait_for_completion, new_tokio_cancellation_token, new_tokio_task_tracker,
